created a utility folder to store the api key
